Finalizando implementacao

+++ src/Lanchonete.Application/ViewModels/LancheViewModel.cs
@@ -5,6 +5,6 @@ public class LancheViewModel
     public Guid Id { get; set; }
     public string Nome { get; set; }
     public string Descricao { get; set; }
-    public double Preco { get; set; }
+    public float Preco { get; set; }
     public bool Status { get; set; }
 }
